+++ blocks/battery.h
@@ -1,2 +1,2 @@
-int batteryu(char *str, int ac);
+size_t batteryu(char *str, int ac);
 void batteryc(int button);
