+++ blocks/calendar.h
@@ -1,2 +1,2 @@
-int calendaru(char *str, int sigval);
+size_t calendaru(char *str, int sigval);
 void calendarc(int button);
