+++ blocks/cputemp.h
@@ -1,2 +1,2 @@
-int cputempu(char *str, int sigval);
+size_t cputempu(char *str, int sigval);
 void cputempc(int button);
