+++ blocks/volume.h
@@ -1,2 +1,2 @@
-int volumeu(char *str, int sigval);
+size_t volumeu(char *str, int sigval);
 void volumec(int button);
